# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "سیستم اشتراک‌گذاری فایل",
        "description": "API برای سیستم اشتراک‌گذاری فایل با Laravel، هر کاربر 5 گیگابایت فضا دارد",
        "contact": {
            "email": "info@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server محلی"
        }
    ],
    "paths": {
...
        "/api/files": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "دریافت لیست فایل‌های کاربر",
                "description": "API: Get list of user files.",
                "operationId": "f7c418d24dac85921d34020fc6e54123",
                "responses": {
                    "200": {
                        "description": "لیست فایل‌های کاربر",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "document.pdf"
                                                    },
                                                    "mime_type": {
                                                        "type": "string",
                                                        "example": "application/pdf"
                                                    },
                                                    "size": {
                                                        "type": "string",
                                                        "example": "2.5 MB"
                                                    },
                                                    "uploaded_at": {
                                                        "type": "string",
                                                        "example": "2 hours ago"
                                                    },
                                                    "shared_with_count": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/files/shared": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "دریافت لیست فایل‌های به اشتراک گذاشته شده با کاربر",
                "description": "API: Get list of files shared with the user.",
                "operationId": "4a4d70241c336e8159fea7ff82411a2f",
                "responses": {
                    "200": {
                        "description": "لیست فایل‌های به اشتراک گذاشته شده با کاربر",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "document.pdf"
                                                    },
                                                    "mime_type": {
                                                        "type": "string",
                                                        "example": "application/pdf"
                                                    },
                                                    "size": {
                                                        "type": "string",
                                                        "example": "2.5 MB"
                                                    },
                                                    "owner": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "shared_at": {
                                                        "type": "string",
                                                        "example": "5 days ago"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}